Combined some services and controllers
